# contracts/openapi/ragforge-api-v1.yaml
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "RAGForge API",
    "version": "1.0.0",
    "description": "Phase 1, Phase 2 and Phase 3 public REST/SSE contract plus frozen Phase 4-G Chunk Studio and Retrieval Playground projections. Phase 1-3 operations retain their planned/contract status; Phase 4-G is a frozen implementation contract and does not claim that the whole API is implemented.",
    "license": {
      "name": "Proprietary project contract"
    }
  },
...
  "servers": [
    {
      "url": "/",
      "description": "The deployment prefixes this contract with the configured origin. API paths include /api/v1."
    }
  ],
...
  "paths": {
...
    "/api/v1/spaces/{spaceId}/sources/uploads": {
      "x-ragforge-implementation-status": "phase-business-loop",
      "parameters": [{"$ref": "#/components/parameters/SpaceId"}, {"$ref": "#/components/parameters/CorrelationId"}],
      "post": {
        "tags": ["ingestion"],
        "operationId": "uploadMarkdownSource",
        "summary": "Upload a Markdown source and enqueue real ingestion",
        "security": [{"cookieAuth": []}, {"serviceToken": []}],
        "parameters": [{"$ref": "#/components/parameters/CorrelationId"}, {"$ref": "#/components/parameters/CsrfToken"}, {"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"type": "object", "required": ["file"], "properties": {"file": {"type": "string", "format": "binary", "description": "UTF-8 Markdown, maximum 10 MiB."}}}}}},
        "responses": {
          "200": {"description": "Upload accepted and represented by a persisted ingestion job", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BusinessUploadResponse"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}, "409": {"$ref": "#/components/responses/Conflict"}, "422": {"$ref": "#/components/responses/UnprocessableEntity"}, "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/api/v1/spaces/{spaceId}/source-documents": {
      "x-ragforge-implementation-status": "phase-business-loop",
      "parameters": [{"$ref": "#/components/parameters/SpaceId"}, {"$ref": "#/components/parameters/CorrelationId"}],
      "get": {"tags": ["ingestion"], "operationId": "listBusinessSourceDocuments", "summary": "List source documents in the authorized space", "security": [{"cookieAuth": []}, {"serviceToken": []}], "parameters": [{"$ref": "#/components/parameters/CorrelationId"}], "responses": {"200": {"description": "Source documents", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/BusinessSourceDocument"}}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}}}
    },
    "/api/v1/spaces/{spaceId}/ingestion-jobs": {
      "x-ragforge-implementation-status": "phase-business-loop",
      "parameters": [{"$ref": "#/components/parameters/SpaceId"}, {"$ref": "#/components/parameters/CorrelationId"}],
      "get": {"tags": ["ingestion"], "operationId": "listBusinessIngestionJobs", "summary": "List ingestion jobs, attempts and pipeline steps", "security": [{"cookieAuth": []}, {"serviceToken": []}], "parameters": [{"$ref": "#/components/parameters/CorrelationId"}], "responses": {"200": {"description": "Ingestion jobs", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/BusinessJobView"}}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}}}
    },
    "/api/v1/spaces/{spaceId}/ingestion-jobs/{jobId}": {
      "x-ragforge-implementation-status": "phase-business-loop",
      "parameters": [{"$ref": "#/components/parameters/SpaceId"}, {"$ref": "#/components/parameters/JobId"}, {"$ref": "#/components/parameters/CorrelationId"}],
      "get": {"tags": ["ingestion"], "operationId": "getBusinessIngestionJob", "summary": "Get one ingestion job with attempt and step evidence", "security": [{"cookieAuth": []}, {"serviceToken": []}], "parameters": [{"$ref": "#/components/parameters/CorrelationId"}], "responses": {"200": {"description": "Ingestion job", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BusinessJobView"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}, "404": {"$ref": "#/components/responses/NotFound"}}}
    },
    "/api/v1/spaces/{spaceId}/document-revisions/{revisionId}/parse-report": {
      "x-ragforge-implementation-status": "phase-business-loop",
      "parameters": [{"$ref": "#/components/parameters/SpaceId"}, {"$ref": "#/components/parameters/RevisionId"}, {"$ref": "#/components/parameters/CorrelationId"}],
      "get": {"tags": ["ingestion"], "operationId": "getBusinessParseReport", "summary": "Get the persisted parse report for a revision", "security": [{"cookieAuth": []}, {"serviceToken": []}], "parameters": [{"$ref": "#/components/parameters/CorrelationId"}], "responses": {"200": {"description": "Parse report", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BusinessParseReport"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}, "404": {"$ref": "#/components/responses/NotFound"}}}
    },
    "/api/v1/spaces/{spaceId}/indexes": {
      "x-ragforge-implementation-status": "phase-business-loop",
      "parameters": [{"$ref": "#/components/parameters/SpaceId"}, {"$ref": "#/components/parameters/CorrelationId"}],
      "get": {"tags": ["indexing"], "operationId": "listBusinessIndexes", "summary": "List real index versions", "security": [{"cookieAuth": []}, {"serviceToken": []}], "parameters": [{"$ref": "#/components/parameters/CorrelationId"}], "responses": {"200": {"description": "Index versions", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/BusinessIndex"}}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}}}
    },
    "/api/v1/spaces/{spaceId}/indexes/active": {
      "x-ragforge-implementation-status": "phase-business-loop",
      "parameters": [{"$ref": "#/components/parameters/SpaceId"}, {"$ref": "#/components/parameters/CorrelationId"}],
      "get": {"tags": ["indexing"], "operationId": "getBusinessActiveIndex", "summary": "Get the active index pointer and server-derived dataset hash", "security": [{"cookieAuth": []}, {"serviceToken": []}], "parameters": [{"$ref": "#/components/parameters/CorrelationId"}], "responses": {"200": {"description": "Active index", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BusinessActiveIndex"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}}}
    },
    "/api/v1/spaces/{spaceId}/indexes/{indexVersionId}/publish": {
      "x-ragforge-implementation-status": "phase-business-loop",
      "parameters": [{"$ref": "#/components/parameters/SpaceId"}, {"$ref": "#/components/parameters/IndexVersionId"}, {"$ref": "#/components/parameters/CorrelationId"}],
      "post": {"tags": ["indexing"], "operationId": "publishBusinessIndex", "summary": "Publish a READY candidate index", "security": [{"cookieAuth": []}, {"serviceToken": []}], "parameters": [{"$ref": "#/components/parameters/CorrelationId"}, {"$ref": "#/components/parameters/CsrfToken"}, {"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"200": {"description": "Active pointer", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ActiveIndexPointer"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}, "409": {"$ref": "#/components/responses/Conflict"}}}
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "RAGFORGE_SESSION",
        "description": "Opaque server-side session cookie. It is HttpOnly, Secure outside local development, and SameSite.",
        "x-ragforge-no-jwt": true
      },
      "serviceToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "service-token",
        "description": "Short-lived, scoped and revocable service token. Tokens are stored as hashes and never logged."
      }
    },
    "parameters": {
      "CorrelationId": {
        "name": "X-Correlation-Id",
        "in": "header",
        "required": false,
        "description": "Client-supplied UUIDv7 for tracing. The server generates one when absent and echoes the effective value.",
        "schema": {"$ref": "#/components/schemas/UuidV7"}
      },
      "CsrfToken": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": true,
        "description": "Required for browser state-changing requests; obtain it from the current session response or a CSRF bootstrap flow.",
        "schema": {"type": "string", "minLength": 32, "maxLength": 256, "pattern": "^[A-Za-z0-9._~-]+$"}
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Stable key for a retried write. The server binds the key to the request hash and refuses a conflicting reuse.",
        "schema": {"type": "string", "minLength": 1, "maxLength": 255, "pattern": "^[A-Za-z0-9._~-]+$"}
      },
      "SpaceId": {
        "name": "spaceId",
        "in": "path",
        "required": true,
        "schema": {"$ref": "#/components/schemas/UuidV7"}
      },
      "JobId": {"name": "jobId", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/UuidV7"}},
      "RevisionId": {"name": "revisionId", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/UuidV7"}},
      "IndexVersionId": {"name": "indexVersionId", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/UuidV7"}},
...
    },
...
    "schemas": {
      "UuidV7": {
        "type": "string",
        "format": "uuidv7",
        "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$",
        "description": "RFC 9562 UUID version 7 with the RFC variant bits."
      },
...
      "ErrorClass": {"type": "string", "enum": ["AUTHENTICATION", "RATE_LIMIT", "QUOTA", "MODEL_NOT_FOUND", "CONTEXT_OVERFLOW", "CONTENT_POLICY", "TIMEOUT", "UNAVAILABLE", "UNSUPPORTED_CAPABILITY", "INVALID_RESPONSE", "CANCELLED", "SPACE_EGRESS_DENIED", "IDEMPOTENCY_CONFLICT"]},
...
      "BusinessUploadResponse": {"type": "object", "additionalProperties": false, "required": ["spaceId", "sourceId", "sourceDocumentId", "documentRevisionId", "jobId", "attemptId", "displayName", "status"], "properties": {"spaceId": {"$ref": "#/components/schemas/UuidV7"}, "sourceId": {"$ref": "#/components/schemas/UuidV7"}, "sourceDocumentId": {"$ref": "#/components/schemas/UuidV7"}, "documentRevisionId": {"$ref": "#/components/schemas/UuidV7"}, "jobId": {"$ref": "#/components/schemas/UuidV7"}, "attemptId": {"$ref": "#/components/schemas/UuidV7"}, "displayName": {"type": "string"}, "status": {"type": "string", "enum": ["REQUESTED", "RUNNING", "SUCCEEDED", "FAILED", "RETRY_SCHEDULED", "DLQ"]}}},
      "BusinessSourceDocument": {"type": "object", "additionalProperties": false, "required": ["id", "spaceId", "sourceId", "stableSourceObjectId", "canonicalSourcePath", "basename", "versionNo", "state"], "properties": {"id": {"$ref": "#/components/schemas/UuidV7"}, "spaceId": {"$ref": "#/components/schemas/UuidV7"}, "sourceId": {"$ref": "#/components/schemas/UuidV7"}, "stableSourceObjectId": {"type": "string"}, "canonicalSourcePath": {"type": "string"}, "basename": {"type": "string"}, "versionNo": {"type": "integer", "minimum": 1}, "state": {"type": "string"}, "activeRevisionId": {"oneOf": [{"$ref": "#/components/schemas/UuidV7"}, {"type": "null"}]}}},
      "BusinessJobView": {"type": "object", "additionalProperties": false, "required": ["job", "attempts", "steps"], "properties": {"job": {"type": "object"}, "attempts": {"type": "array", "items": {"type": "object"}}, "steps": {"type": "array", "items": {"type": "object"}}}},
      "BusinessParseReport": {"type": "object", "additionalProperties": false, "required": ["parseReportId", "documentRevisionId", "status", "mediaType", "pageCount", "characterCount", "tokenCount", "parserName", "parserVersion", "durationMs", "createdAt"], "properties": {"parseReportId": {"$ref": "#/components/schemas/UuidV7"}, "documentRevisionId": {"$ref": "#/components/schemas/UuidV7"}, "status": {"type": "string"}, "mediaType": {"type": "string"}, "pageCount": {"type": "integer"}, "characterCount": {"type": "integer"}, "tokenCount": {"type": "integer"}, "parserName": {"type": "string"}, "parserVersion": {"type": "string"}, "durationMs": {"type": "integer"}, "warnings": {"type": "string"}, "errors": {"type": "string"}, "extractedTextArtifactId": {"oneOf": [{"$ref": "#/components/schemas/UuidV7"}, {"type": "null"}]}, "ocrStatus": {"type": "string"}, "ocrEngine": {"type": ["string", "null"]}, "ocrEngineVersion": {"type": ["string", "null"]}, "createdAt": {"type": "string", "format": "date-time"}}},
      "BusinessIndex": {"type": "object", "additionalProperties": false, "required": ["indexVersionId", "spaceId", "versionNo", "state", "candidateCollection", "embeddingProfileVersion", "chunkingStrategyVersion", "documentRevisionCount", "childChunkCount", "createdAt"], "properties": {"indexVersionId": {"$ref": "#/components/schemas/UuidV7"}, "spaceId": {"$ref": "#/components/schemas/UuidV7"}, "versionNo": {"type": "integer"}, "state": {"type": "string"}, "candidateCollection": {"type": "string"}, "embeddingProfileVersion": {"type": "string"}, "chunkingStrategyVersion": {"type": "string"}, "documentRevisionCount": {"type": "integer"}, "childChunkCount": {"type": "integer"}, "validationVectorDimension": {"type": ["integer", "null"]}, "sampleRetrievalPassed": {"type": ["boolean", "null"]}, "spaceFilterPassed": {"type": ["boolean", "null"]}, "activatedAt": {"type": ["string", "null"], "format": "date-time"}, "createdAt": {"type": "string", "format": "date-time"}}},
      "BusinessActiveIndex": {"type": "object", "additionalProperties": false, "required": ["pointer", "index", "datasetHash"], "properties": {"pointer": {"$ref": "#/components/schemas/ActiveIndexPointer"}, "index": {"oneOf": [{"$ref": "#/components/schemas/BusinessIndex"}, {"type": "null"}]}, "datasetHash": {"oneOf": [{"type": "string", "pattern": "^[0-9a-f]{64}$"}, {"type": "null"}]}}},
      "ActiveIndexPointer": {"type": "object", "additionalProperties": false, "required": ["id", "spaceId", "activeIndexVersionId", "versionNo", "updatedAt"], "properties": {"id": {"$ref": "#/components/schemas/UuidV7"}, "spaceId": {"$ref": "#/components/schemas/UuidV7"}, "activeIndexVersionId": {"$ref": "#/components/schemas/UuidV7"}, "previousIndexVersionId": {"oneOf": [{"$ref": "#/components/schemas/UuidV7"}, {"type": "null"}]}, "versionNo": {"type": "integer"}, "updatedAt": {"type": "string", "format": "date-time"}}},
...
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["field", "code", "message"],
        "properties": {
          "field": {"type": "string", "minLength": 1, "maxLength": 200},
          "code": {"type": "string", "pattern": "^[A-Z][A-Z0-9_]{1,63}$"},
          "message": {"type": "string", "minLength": 1, "maxLength": 500}
        }
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": false,
        "required": ["type", "title", "status", "detail", "instance", "code", "correlationId"],
        "properties": {
          "type": {"type": "string", "format": "uri-reference"},
          "title": {"type": "string", "minLength": 1},
          "status": {"type": "integer", "minimum": 400, "maximum": 599},
          "detail": {"type": "string", "minLength": 1, "maxLength": 2000},
          "instance": {"type": "string", "format": "uri-reference"},
          "code": {"type": "string", "pattern": "^[A-Z][A-Z0-9_]{1,63}$"},
          "correlationId": {"$ref": "#/components/schemas/UuidV7"},
          "fieldErrors": {"type": "array", "maxItems": 50, "items": {"$ref": "#/components/schemas/FieldError"}},
          "errorClass": {"$ref": "#/components/schemas/ErrorClass"},
          "retryable": {"type": "boolean"},
          "retryAfterSeconds": {"type": "integer", "minimum": 1, "maximum": 3600}
        },
        "example": {
          "type": "https://errors.ragforge.local/problems/space-access-denied",
          "title": "Space access denied",
          "status": 403,
          "detail": "The principal is not a member of this space.",
          "instance": "/api/v1/spaces/0190f5c2-7c1e-7abc-8def-1234567890ab",
          "code": "SPACE_ACCESS_DENIED",
          "correlationId": "0190f5c2-7c1e-7abc-8def-1234567890ab"
        }
      }
    },
    "responses": {
...
      "Unauthorized": {"description": "Authentication required or credentials invalid", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetails"}}}},
      "Forbidden": {"description": "Permission or CSRF check failed", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetails"}}}},
      "NotFound": {"description": "Resource not found in the authorized space", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetails"}}}},
      "Conflict": {"description": "Duplicate or idempotency/request-hash conflict", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetails"}}}},
...
      "UnprocessableEntity": {"description": "Request failed field validation", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetails"}}}},
...
      ,"ServiceUnavailable": {"description": "Required local infrastructure is unavailable", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetails"}}}}
    }
  }
}